\name{corneta}
\alias{corneta}
\docType{data}
\title{Corneta Outcrop}
\description{
  Pollen counts from a stratigraphic section at the Corneta outcrop. This is to illustrate how to read in `raw' data in an arbitrary format, so it has been left entirely as character data.
}
\usage{data(corneta)}
\format{
  A data frame with 13 observations (rows) of 169 variables (columns).
  \describe{
    The rows represent depths and the columns taxa.
  }
}

\details{
  Not yet available.
}

\source{
  Silane da Silva
}

\references{}

\examples{
  data(corneta)
  str(corneta)
}

\keyword{datasets}