\name{gluesheets}
\alias{gluesheets}
\title{Glues tab-delimited spreadsheets together for further analysis}
\description{
Glues tab-delimited spreadsheets of species (the columns) by stratigraphic depths (the rows) together in such a way that the output is correctly formatted for analysis of `unitary associations' in the PAST software package.}

\usage{
gluesheets(PATH = NULL, out.file = 'fileout.txt')
}

\arguments{
  \item{PATH}{a string giving the file path to the directory in which the sheets to be glued together are located.}
  \item{out.file}{the name of the file to which to write the output.}
}

\details{
  Ask Silane da Silva....
}

\value{}
\references{}

\author{Walton Green}

\note{This is legacy code that will probably be removed soon....}
%\section{Warning }{....}

\seealso{\code{\link{plot.strat.column}}}

\examples{
  ## No examples
}

\keyword{manip}