\name{nonuser}
\alias{adding}
\alias{getsse}
\alias{cpredict}
\alias{mergelist}
\alias{rbind.all}
\title{Non user-level subroutines}
\description{
Subroutines called by other functions; none are intended to be called directly by the user.}

%\usage{}
%\arguments{}
%\details{}
%\value{}
%\references{}
\author{Walton Green}
%\note{}

%Make other sections like Warning with \section{Warning }{....}

\seealso{\code{\link{plot.strat.column}},
         \code{\link{stratigraph}}}

%\examples{}
\keyword{misc}