\name{plain}
\alias{plain}
\alias{striped}
\alias{stippled}
\docType{data}
\title{`plain' example section}
\description{
  Three very simple example sections used as examples by Edwards (1984).
}
\usage{data(plain)}
\format{
  Data frames with taxa in columns (except the first column, which has the depth of each sample) and samples in rows.
}
\details{}

\source{
  Edwards, L. E. (1984) Insights on why graphic correlation (Shaw's
Method) works. Journal of Geology 92:583--597.
}

\references{}
\examples{
data(plain)
str(plain)
}
\keyword{datasets}
