\name{sI}
\alias{sI}
\alias{sII}
\alias{sIII}
\alias{sIV}
\alias{sV}
\alias{sVI}
\docType{data}
\title{Section I}
\description{
    Six synthetic sections used as examples by Edwards (1984).
}
\usage{data(sI)}
\format{
  Data frames with taxa in columns (except the first column, which has the depth of each sample) and samples in rows.
}
\details{}
\source{

Edwards, L. E. (1984) `Insights on why graphic correlation (Shaw's Method) works' Journal of Geology 92:583--597.

}
\references{}
\examples{
data(sI)
str(sI)
}
\keyword{datasets}