\name{stratigraph-package}
\alias{stratgraph-package}
\alias{stratigraph}
\docType{package}
\title{
Plots stratigraphic data
}
\description{
A collection of tools for plotting and analyzing paleontological and geological data distributed through through time in stratigraphic cores or sections. Includes some miscellaneous functions for handling other kinds of palaeontological and paleoecological data.
}
\details{
\tabular{ll}{
  Package: \tab stratigraph\cr
  Type: \tab Package\cr
  Version: \tab 0.6\cr
  Date: \tab 2010-03-30\cr
  License: \tab GPL v. 3 or later copyleft\cr
  }

Package overview; still in beta.
Includes functions:
  \code{\link{writeTilia}}\cr
  \code{\link{writeGPDascii}}\cr
  \code{\link{readTilia}}\cr
  \code{\link{readGPDascii}}\cr
  \code{\link{gluesheets}}\cr
\cr
  \code{\link{a.datums}}\cr
  \code{\link{fads}}\cr
  \code{\link{lads}}\cr
  \code{\link{fill.ranges}}\cr
  \code{\link{addpart}}\cr
  \code{\link{oe.rates}}\cr
\cr
  \code{\link{plot.strat.column}}\cr
  \code{\link{strat.column}}\cr
  \code{\link{as.strat.column}}\cr
  \code{\link{is.strat.column}}\cr
  \code{\link{summary.strat.column}}\cr

  \code{\link{dloc}}\cr
  \code{\link{dcloc}}\cr

and data:
\code{\link{plain}}\cr
\code{\link{striped}}\cr
\code{\link{stippled}}\cr
\code{\link{sI}}\cr
\code{\link{sII}}\cr
\code{\link{sIII}}\cr
\code{\link{sIV}}\cr
\code{\link{sV}}\cr
\code{\link{sVI}}\cr
\code{\link{rogerslk}}\cr
\code{\link{mohawk}}\cr
\code{\link{corneta}}\cr
\code{\link{cerrejon}}\cr
}

\author{
Walton A. Green, Carlos Jaramillo, Scott Wing

Maintainer: Walton A. Green <wagreen@bricol.net>
}

\references{
Shaw, A. B. (1964) Time in Stratigraphy. New York: McGraw-Hill.

Edwards, L. E. (1984) Insights on why graphic correlation (Shaw's
Method) works. Journal of Geology 92:583--597.
}

\keyword{package}
%Optionally other keywords from doc/KEYWORDS, one per line

\seealso{
\code{\link{plot.strat.column}}
}

\examples{
  cat('Put some examples here')
}